+++ activerecord/lib/active_record/connection_adapters/sqlite3/schema_creation.rb
@@ -5,6 +5,12 @@ module ActiveRecord
     module SQLite3
       class SchemaCreation < AbstractAdapter::SchemaCreation # :nodoc:
         private
+          def visit_AddColumnDefinition(o)
+            sql = "ADD #{accept(o.column)}"
+            sql += " REFERENCES #{o.column.references}" if o.column.references
+            sql.dup
+          end
+
           def add_column_options!(sql, options)
             if options[:collation]
               sql << " COLLATE \"#{options[:collation]}\""